Quantified memory, predict full model will be ~200 mb
Trying to decide whether to re-do my current memory intensive
approach I used fil-profile and measured memory footprint.

Basically, my current approach is not ideal, but it'll be fine and I won't
need to bother improving it for some time.

Predicted that a model with 1000 vars and 870 constraints will take
168.27 MiB using linear regression.

Results are in memory.txt